Add getUserRides() to a rideListServiceStub
The 10 failing tests in user-profile.component.spec.ts needed that property in the stub. The unit test
"Editing a user calls UserListService.editUser" failed because
"TypeError: Cannot read property 'userId' of undefined"

The other 9 tests would likely move onto that error. I think that issue
is not about oauth but about how the tests are handled in general. For the time being, this commit makes note of that so that testers can move onto other oauth related tests.

#4