Add additional info about client tests
Some tests need the new fields. And regarding the window['gapi'] part,
we are not sure if it is actually needed in what it is used in, but
before merging into master we will figure that out. It was something
most if not all tests had at some point when we were trying various
things. It may be the case that it is only needed in select tests.

#4